requirements.txt
PyYAML>=6.0
reportlab>=4.0
mistune>=3.0
